動きます

+++ etrobo_honban2/CompetitionRunning/RunningController.cpp
@@ -10,9 +10,9 @@ RunningController::RunningController(DeviceValueGet* DeviceValueGet,
 }
 
 //メソッド： void 走行する（現区間の番号）
-int8_t RunningController::RunningExecute(int now_section){
+double RunningController::RunningExecute(int now_section){
 	DeviceValue dv;
-	int8_t turn;
+	double turn;
 
 	//現在のデバイス値を持ってくる
 	dv = mDeviceValueGet->DeviceValueGetter();
